Make Addr a u64
So far Addr was defined to be a usize. That is in line with Rust
specifying that usize has the size of a pointer. However, once we start
using the library in a cross compilation context (e.g., using it to
symbolize parts of a dump captured on a different machine), the local
pointer size has admittedly little meaning.
With this change we define Addr to be of type u64, which is in-line with
the representation of many of the formats that we support. Doing so
makes us more independent of the actual pointer size as it basically is
meant as an upper limit on current sizes (that may not be entirely true
with research architectures such as CHERI, but it is still a reasonable
and arguably better approximation for what we expect to support).

+++ src/dwarf/resolver.rs
@@ -105,7 +105,7 @@ impl DwarfResolver {
         //       unnecessary. Consider removing it or moving it higher
         //       in the call chain.
         let code_info = if self.line_number_info {
-            if let Some(direct_location) = self.units.find_location(addr as u64)? {
+            if let Some(direct_location) = self.units.find_location(addr)? {
                 let Location {
                     dir,
                     file,
@@ -121,7 +121,7 @@ impl DwarfResolver {
                 };
 
                 let inlined = if inlined_fns {
-                    if let Some(inline_stack) = self.units.find_inlined_functions(addr as u64)? {
+                    if let Some(inline_stack) = self.units.find_inlined_functions(addr)? {
                         let mut inlined = Vec::with_capacity(inline_stack.len());
                         for result in inline_stack {
                             let (name, location) = result?;
@@ -186,17 +186,14 @@ impl DwarfResolver {
             ))
         }
 
-        let result = self.units.find_function(addr as u64)?;
+        let result = self.units.find_function(addr)?;
         if let Some((function, language)) = result {
             let name = function
                 .name
                 .map(|name| name.to_string())
                 .transpose()?
                 .unwrap_or("");
-            let addr = function
-                .range
-                .map(|range| range.begin as usize)
-                .unwrap_or(0);
+            let addr = function.range.map(|range| range.begin).unwrap_or(0);
             let size = function
                 .range
                 .map(|range| usize::try_from(range.end - range.begin).unwrap_or(usize::MAX));
